fixed race condition on multiple threads updated to c++ types

+++ src/board.h
@@ -13,37 +13,37 @@ class Board
 public:
 	Board();
 	~Board();
-	void init(Component** components, Link* links, unsigned int componentCount, unsigned int linkCount);
-	unsigned int getNextComponentIndex();
+	void init(Component** components, Link* links, size_t componentCount, size_t linkCount);
+	size_t getNextComponentIndex();
 	size_t componentCount = 0;
 	size_t linkCount = 0;
 	bool* linkStates = nullptr;
 	enum State { Uninitialized, Stopped, Running, Stopping };
-	unsigned int getThreadCount() const;
+	size_t getThreadCount() const;
 	Component** getComponents() const;
 	Link* getLinks() const;
 	State getCurrentState() const;
-	unsigned long long int getCurrentTick() const;
-	unsigned long long currentSpeed = 0;
+	uint_fast64_t getCurrentTick() const;
+	uint_fast64_t currentSpeed = 0;
 	Events::Event<> tickEvent;
 	void stop();
-	void start(unsigned long long cyclesLeft, unsigned long ms, unsigned int threadCount, bool synchronized = false);
+	void start(uint_fast64_t cyclesLeft, uint_fast32_t ms, size_t threadCount, bool synchronized = false);
 	FastStackAtomic<Link*>* readBuffer = new FastStackAtomic<Link*>();
 	FastStackAtomic<Link*>* writeBuffer = new FastStackAtomic<Link*>();
 
 private:
-	unsigned int threadCount = 0;
+	size_t threadCount = 0;
 	Component** components = nullptr;
 	Link* links = nullptr;
 	State currentState = Board::Uninitialized;
-	unsigned long long int tick = 0;
+	uint_fast64_t tick = 0;
 	std::thread** threads = nullptr;
 	SpinlockBarrier* barrier = nullptr;
 	std::chrono::high_resolution_clock::time_point lastCapture;
 	std::chrono::high_resolution_clock::time_point started;
-	unsigned long long int lastCaptureTick = 0;
-	unsigned int componentIndex = 0;
-	unsigned long long cyclesLeft = 0;
-	unsigned long long timeout = 0;
+	uint_fast64_t lastCaptureTick = 0;
+	size_t componentIndex = 0;
+	uint_fast64_t cyclesLeft = 0;
+	uint_fast64_t timeout = 0;
 };
