+++ src/component.h
@@ -28,12 +28,12 @@ public:
 		return outputs;
 	}
 
-	unsigned int getInputCount() const
+	size_t getInputCount() const
 	{
 		return inputCount;
 	};
 
-	unsigned int getOutputCount() const
+	size_t getOutputCount() const
 	{
 		return outputCount;
 	};
@@ -42,26 +42,26 @@ protected:
 	Board* board;
 	Input* inputs;
 	Output* outputs;
-	unsigned int inputCount;
-	unsigned int outputCount;
+	size_t inputCount;
+	size_t outputCount;
 
-	Component(Board* board, Link** inputs, Link** outputs, const unsigned int inputCount, const unsigned int outputCount) :
+	Component(Board* board, Link** inputs, Link** outputs, const size_t inputCount, const size_t outputCount) :
 		board(board),
 		inputCount(inputCount),
 		outputCount(outputCount)
 	{
 		this->inputs = new Input[inputCount];
-		for (unsigned int i = 0; i < inputCount; i++) {
+		for (size_t i = 0; i < inputCount; i++) {
 			new (&this->inputs[i]) Input(this, inputs[i]);
 		}
 
 		this->outputs = new Output[outputCount];
-		for (unsigned int i = 0; i < outputCount; i++) {
+		for (size_t i = 0; i < outputCount; i++) {
 			new (&this->outputs[i]) Output(this, outputs[i]);
 		}
 
 
-		for (unsigned int i = 0; i < inputCount; i++) {
+		for (size_t i = 0; i < inputCount; i++) {
 			auto** newInputs = new Input* [inputs[i]->inputCount + 1u];
 			std::memcpy(newInputs, inputs[i]->inputs, inputs[i]->inputCount * sizeof(Input*));
 			newInputs[inputs[i]->inputCount] = &this->inputs[i];
@@ -70,7 +70,7 @@ protected:
 			inputs[i]->inputCount++;
 		}
 
-		for (unsigned int i = 0; i < outputCount; i++) {
+		for (size_t i = 0; i < outputCount; i++) {
 			auto** newOutputs = new Output * [outputs[i]->outputCount + 1];
 			std::memcpy(newOutputs, outputs[i]->outputs, outputs[i]->outputCount * sizeof(Output*));
 			newOutputs[outputs[i]->outputCount] = &this->outputs[i];
@@ -80,7 +80,7 @@ protected:
 		}
 	}
 
-	Component(Board* board, Input* inputs, Output* outputs, const unsigned int inputCount, const unsigned int outputCount) :
+	Component(Board* board, Input* inputs, Output* outputs, const size_t inputCount, const size_t outputCount) :
 		board(board),
 		inputs(inputs),
 		outputs(outputs),
